record and send to firebase
