+++ src/App.js
@@ -1,17 +1,17 @@
 import React, { useState } from "react";
 import "./App.css";
 
-import AudioRecorder from './components/AudioRecorder/AudioRecorder';
 import Login from './components/Authentication/login';
 import Signup from './components/Authentication/signup';
-import SendAudio from "./components/AudioRecorder/sendAudioToFirebase";
+import AudioSender from "./components/AudioRecorder/AudioSender";
 
 function App() {
   return (
     <div className="App">
-      {/* <Signup /> */}
+      {/* <Signup /> */}  
       {/* <Login /> */}
-      <AudioRecorder />
+      {/* <AudioRecorder /> */}
+      <AudioSender/>
       {/* <SendAudio /> */}
     </div>
   );
