--- src/components/AudioRecorder/styles.css
@@ -1,24 +1,0 @@
-
-
-.record-button{
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, 100%);
-
-    justify-content: center;
-    align-items: center;
-
-
-    button{
-        
-        display: hidden;
-        background: transparent;
-        border: transparent;
-
-        img{
-            width: 100px;
-            height: 100px;
-        }
-    }
-}
